much better physics

+++ Assets/Scripts/GameManager.cs
@@ -11,7 +11,7 @@ public class GameManager : MonoBehaviour
     public GameState _gameState;
     public TextMeshProUGUI tp;
     public float originalSpeed;
-    public int Score;
+    public float Score;
     private void Awake()
     {
         originalSpeed = CharacterMovement.instance.speed;
